ai_service/keyword_rewrite/requirements.txt
# Web Server
flask==2.3.3
flask_cors==4.0.0

# Scraping và Requests
requests>=2.28.0,<3.0.0
beautifulsoup4==4.12.2
trafilatura==1.6.1
selenium==4.10.0
webdriver-manager==3.8.5
charset-normalizer>=3.1.0
urllib3<2.0.0
bs4
lxml[html_clean]

# AI Services
google-generativeai==0.3.1

# Utilities
python-dotenv==0.21.1 
